Split options packet into multiple packets
Due to the huge amount of options, the SyncCustomSettings packet is
getting extremely large. This patch takes two measures to correct this

1. By using StartRpcImmediately it'll not lump in this packet with
around ~100 bytes of other data. This moves the issue 20 options down
the road.

2. By splitting the packet in multiple chunks of 1000 bytes, every
packet stays well beloy the ~1472 byte limit. This prevents the issue
from appearing again. This change needs the first change though,
otherwise both chunks will be lumped in the same packet.